chore: move script and style outside gatsby-config.js
