+++ site/global.less
@@ -1,0 +1,24 @@
+.mobile-container {
+  width: 395px !important;
+  height: 590px;
+  background-image: url(https://gw.alipayobjects.com/zos/rmsportal/JAxcjwaQvIsAWvXUrjbY.png);
+  background-size: 100%;
+  background-repeat: no-repeat;
+  overflow: hidden;
+}
+
+.mobile-header {
+  height: 64px;
+  overflow: hidden;
+}
+
+.mobile-content {
+  margin: 15px 10px;
+  background-color: #fff;
+
+  canvas {
+    display: block;
+    width: 375px;
+    height: 260px;
+  }
+}
